Fix movement synchronization issue
Using a global 'frame_count' to calculate the interval for the 'move_decision' function being called would cause every organism, occasionally, to make a new movement decision at the same time, resulting in them 'vibrating' together. To fix this, I added a 'last_decision' value to the Organism class, which is used in the calculation instead of 'frame_count'.